Make CharacterPartMovement a system
