--- Assets/Scripts/Field/Character/CharacterPartMovement.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: b98ba5006472427a8085777a8cc301ba
-timeCreated: 1655102699
